@@ -1,20 +1,14 @@
 #
-#  Copyright (c) 2023 Composiv.ai
+# Copyright (c) 2023 Composiv.ai
 #
-# All rights reserved. This program and the accompanying materials
-# are made available under the terms of the Eclipse Public License v2.0
-# and Eclipse Distribution License v1.0 which accompany this distribution.
+# This program and the accompanying materials are made available under the
+# terms of the Eclipse Public License 2.0 which is available at
+# http://www.eclipse.org/legal/epl-2.0.
 #
-# Licensed under the  Eclipse Public License v2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-# The Eclipse Public License is available at
-#    http://www.eclipse.org/legal/epl-v20.html
-# and the Eclipse Distribution License is available at
-#   http://www.eclipse.org/org/documents/edl-v10.php.
+# SPDX-License-Identifier: EPL-2.0
 #
 # Contributors:
-#    Composiv.ai - initial API and implementation
+#   Composiv.ai - initial API and implementation
 #
 
 """
@@ -30,29 +24,28 @@ import base64
 import json
 import signal
 import threading
-from typing import Any, Dict, List, Optional, Tuple
+from typing import Any
 
 # Third-party imports
 import rclpy
 from muto_msgs.msg import MutoAction
-from rclpy.node import Node
 from std_srvs.srv import Trigger
+from symphony_sdk import (
+    ComparisonPack,
+    ComponentResultSpec,
+    ComponentSpec,
+    State,
+    SummarySpec,
+    SummaryState,
+    SymphonyAPI,
+    SymphonyAPIError,
+    TargetResultSpec,
+    to_dict,
+)
 
 # Local imports
 from ..config import ConfigurationManager
 from ..interfaces import BaseNode
-from symphony_sdk import (
-    SymphonyAPI,
-    SymphonyAPIError,
-    ComponentSpec,
-    ComparisonPack,
-    ComponentResultSpec,
-    SummarySpec,
-    SummaryState,
-    TargetResultSpec,
-    State,
-    to_dict,
-)
 from .provider_base import SymphonyProvider
 from .symphony_broker import MQTTBroker
 
@@ -60,74 +53,64 @@ from .symphony_broker import MQTTBroker
 class MutoSymphonyProvider(BaseNode, SymphonyProvider):
     """
     Symphony provider integrating with Muto Agent MQTT infrastructure.
-    
+
     This provider implements the Symphony provider interface and manages
     ROS 2 components through Muto's existing MQTT communication layer.
     """
-    
+
     def __init__(
-        self, 
-        config_manager: Optional[ConfigurationManager] = None, 
-        node_name: str = 'muto_symphony_provider'
+        self,
+        config_manager: ConfigurationManager | None = None,
+        node_name: str = "muto_symphony_provider",
     ):
         """
         Initialize the Muto Symphony Provider.
-        
+
         Args:
             config_manager: Optional configuration manager.
             node_name: ROS 2 node name.
         """
         super().__init__(node_name)
-        
+
         # Configuration
         self._config_manager = config_manager or ConfigurationManager(self)
         self._config = self._config_manager.load_config()
-        
+
         # MQTT Broker for Symphony communication
-        self._mqtt_broker: Optional[MQTTBroker] = None
-        
+        self._mqtt_broker: MQTTBroker | None = None
+
         # Symphony API client
-        self._api_client: Optional[SymphonyAPI] = None
-        
+        self._api_client: SymphonyAPI | None = None
+
         # Lifecycle management
         self._shutdown_event = threading.Event()
         self._running = False
-        
+
         # Setup logging
         self.logger = self.get_logger()
-        
+
         # Setup stack publisher
         topics = self._config.topics
-        self.stack_publisher = self.create_publisher(
-            MutoAction, topics.stack_topic, 10
-        )
+        self.stack_publisher = self.create_publisher(MutoAction, topics.stack_topic, 10)
 
-        self._component_registry: Dict[str, Dict[str, Any]] = {}
-    
+        self._component_registry: dict[str, dict[str, Any]] = {}
 
     def _do_initialize(self) -> None:
-    
+
         # Create ROS services for Symphony registration
-        self._register_service = self.create_service(
-            Trigger,
-            'symphony_register_target',
-            self._register_target_service
-        )
-        
+        self._register_service = self.create_service(Trigger, "symphony_register_target", self._register_target_service)
+
         self._unregister_service = self.create_service(
-            Trigger,
-            'symphony_unregister_target', 
-            self._unregister_target_service
+            Trigger, "symphony_unregister_target", self._unregister_target_service
         )
-        
+
         # Initialize Symphony API client
         self._init_symphony_api()
-        
+
         # Initialize MQTT Broker if enabled
         if self._config.symphony.enabled:
             self._init_mqtt_broker()
-            
-    
+
     def _init_symphony_api(self) -> None:
         """Initialize Symphony API client."""
         try:
@@ -137,11 +120,11 @@ class MutoSymphonyProvider(BaseNode, SymphonyProvider):
                 username=symphony.mqtt.user,
                 password=symphony.mqtt.password,
                 timeout=30.0,
-                logger=self.logger
+                logger=self.logger,
             )
-            
+
             self.logger.info("Symphony API client initialized")
-            
+
         except Exception as e:
             self.logger.error(f"Failed to initialize Symphony API client: {e}")
             self._api_client = None
@@ -152,30 +135,26 @@ class MutoSymphonyProvider(BaseNode, SymphonyProvider):
             # Initialize the MQTTBroker with this provider as the plugin
             self._mqtt_broker = MQTTBroker(
                 plugin=self,  # This provider implements the required methods
-                node=self,    # ROS node
-                config=self._config  # Configuration
+                node=self,  # ROS node
+                config=self._config,  # Configuration
             )
             self._mqtt_broker.connect()
-             
+
         except Exception as e:
             self.logger.error(f"Failed to initialize MQTT Broker: {e}")
             self._mqtt_broker = None
 
-
-
-
-
-    def _authenticate_symphony_api(self) -> Optional[str]:
+    def _authenticate_symphony_api(self) -> str | None:
         """
         Authenticate with Symphony API and return access token.
-        
+
         Returns:
             Access token if successful, None if failed.
         """
         if not self._api_client:
             self.logger.error("Symphony API client not initialized")
             return None
-            
+
         try:
             return self._api_client.authenticate()
         except SymphonyAPIError as e:
@@ -185,14 +164,14 @@ class MutoSymphonyProvider(BaseNode, SymphonyProvider):
     def register_target(self) -> bool:
         """
         Register this Muto agent as a target in Symphony.
-        
+
         Returns:
             True if registration successful, False otherwise.
         """
         if not self._api_client:
             self.logger.error("Symphony API client not initialized")
             return False
-            
+
         symphony = self._config.symphony
         try:
             # Build target registration payload
@@ -211,18 +190,18 @@ class MutoSymphonyProvider(BaseNode, SymphonyProvider):
                                     "clientID": symphony.client_id,
                                     "requestTopic": f"{symphony.topic_prefix}/{symphony.request_topic}",
                                     "responseTopic": f"{symphony.topic_prefix}/{symphony.response_topic}",
-                                    "timeoutSeconds": symphony.timeout_seconds
-                                }
+                                    "timeoutSeconds": symphony.timeout_seconds,
+                                },
                             }
                         ]
                     }
-                ]
+                ],
             }
-            
+
             # Register target using API client
             self._api_client.register_target(symphony.target, target_payload)
             return True
-            
+
         except SymphonyAPIError as e:
             self.logger.error(f"Failed to register target with Symphony: {e}")
             return False
@@ -233,20 +212,20 @@ class MutoSymphonyProvider(BaseNode, SymphonyProvider):
     def unregister_target(self) -> bool:
         """
         Unregister this Muto agent target from Symphony.
-        
+
         Returns:
             True if unregistration successful, False otherwise.
         """
         if not self._api_client:
             self.logger.error("Symphony API client not initialized")
             return False
-            
+
         symphony = self._config.symphony
         try:
             # Unregister target using API client
             self._api_client.unregister_target(symphony.target, direct=True)
             return True
-            
+
         except SymphonyAPIError as e:
             self.logger.error(f"Failed to unregister target from Symphony: {e}")
             return False
@@ -272,18 +251,14 @@ class MutoSymphonyProvider(BaseNode, SymphonyProvider):
         except Exception as e:
             self.logger.error(f"Error in auto-registration: {e}")
 
-    def _register_target_service(
-        self, 
-        request: Trigger.Request, 
-        response: Trigger.Response
-    ) -> Trigger.Response:
+    def _register_target_service(self, request: Trigger.Request, response: Trigger.Response) -> Trigger.Response:
         """
         ROS service callback to register target with Symphony.
-        
+
         Args:
             request: Service request.
             response: Service response.
-            
+
         Returns:
             Service response with success status and message.
         """
@@ -295,18 +270,14 @@ class MutoSymphonyProvider(BaseNode, SymphonyProvider):
             response.message = f"Failed to register target '{self._config.symphony.target}' with Symphony"
         return response
 
-    def _unregister_target_service(
-        self, 
-        request: Trigger.Request, 
-        response: Trigger.Response
-    ) -> Trigger.Response:
+    def _unregister_target_service(self, request: Trigger.Request, response: Trigger.Response) -> Trigger.Response:
         """
         ROS service callback to unregister target from Symphony.
-        
+
         Args:
             request: Service request.
             response: Service response.
-            
+
         Returns:
             Service response with success status and message
         """
@@ -323,32 +294,26 @@ class MutoSymphonyProvider(BaseNode, SymphonyProvider):
 
     def init_provider(self):
         """Initialize the provider - required by MQTTBroker interface."""
-        self.logger.info(f"Muto Symphony Provider initialized - Target: {self._config.symphony.target}")    
+        self.logger.info(f"Muto Symphony Provider initialized - Target: {self._config.symphony.target}")
         # Auto-register target after initialization
         self.register_target()
-        
-    def apply(
-        self, 
-        metadata: Dict[str, Any], 
-        components: List[ComponentSpec]
-    ) -> str:
+
+    def apply(self, metadata: dict[str, Any], components: list[ComponentSpec]) -> str:
         """
         Apply/deploy components.
-        
+
         Called by MQTTBroker when Symphony sends deployment requests.
-        This method implements the core deployment logic for Symphony 
+        This method implements the core deployment logic for Symphony
         components.
-        
+
         Args:
             metadata: Request metadata.
             components: List of components to deploy.
-            
+
         Returns:
             JSON string containing deployment results.
         """
-        self.logger.info(
-            f"Symphony apply: deploying {len(components)} components"
-        )
+        self.logger.info(f"Symphony apply: deploying {len(components)} components")
 
         # Use summary models
         result = SummarySpec(target_count=1)
@@ -360,9 +325,7 @@ class MutoSymphonyProvider(BaseNode, SymphonyProvider):
 
         for component in components:
             component_name = component.name or "unnamed-component"
-            self.logger.info(
-                f"Deploying component: {component_name} of type {component.type}"
-            )
+            self.logger.info(f"Deploying component: {component_name} of type {component.type}")
 
             component_result = ComponentResultSpec()
 
@@ -372,24 +335,20 @@ class MutoSymphonyProvider(BaseNode, SymphonyProvider):
                 component_result.status = State.UPDATE_FAILED
                 component_result.message = decode_error
                 target_result.component_results[component_name] = component_result
-                self.logger.error(
-                    f"Component {component_name} payload error: {decode_error}"
-                )
+                self.logger.error(f"Component {component_name} payload error: {decode_error}")
                 continue
 
             publish_method = self._resolve_component_method(component, default="apply")
             published = self._publish_stack_action(
                 method=publish_method,
                 payload=stack_payload,
-                context=component.metadata.get("context", "") if component.metadata else ""
+                context=component.metadata.get("context", "") if component.metadata else "",
             )
 
             if not published:
                 failures += 1
                 component_result.status = State.MQTT_PUBLISH_FAILED
-                component_result.message = (
-                    f"Failed to publish apply action for component {component_name}"
-                )
+                component_result.message = f"Failed to publish apply action for component {component_name}"
                 target_result.component_results[component_name] = component_result
                 self.logger.error(component_result.message)
                 # Do NOT update _component_registry on failed publish
@@ -398,9 +357,7 @@ class MutoSymphonyProvider(BaseNode, SymphonyProvider):
             # Only update registry after confirmed successful publish
             successes += 1
             component_result.status = State.UPDATED
-            component_result.message = (
-                f"Apply action published for component {component_name}"
-            )
+            component_result.message = f"Apply action published for component {component_name}"
             target_result.component_results[component_name] = component_result
 
             # Registry mutation: only performed after successful stack action publish
@@ -414,9 +371,7 @@ class MutoSymphonyProvider(BaseNode, SymphonyProvider):
 
         target_result.status = "OK" if failures == 0 else "FAILED"
         if failures:
-            target_result.message = (
-                f"{failures} component(s) failed during apply"
-            )
+            target_result.message = f"{failures} component(s) failed during apply"
 
         result.success_count = successes
         result.current_deployed = successes
@@ -429,28 +384,22 @@ class MutoSymphonyProvider(BaseNode, SymphonyProvider):
 
         return json.dumps(result.to_dict(), indent=2)
 
-    def remove(
-        self, 
-        metadata: Dict[str, Any], 
-        components: List[ComponentSpec]
-    ) -> str:
+    def remove(self, metadata: dict[str, Any], components: list[ComponentSpec]) -> str:
         """
         Remove components.
-        
+
         Called by MQTTBroker when Symphony sends removal requests.
         This method implements the core removal logic for Symphony components.
-        
+
         Args:
             metadata: Request metadata.
             components: List of components to remove.
-            
+
         Returns:
             JSON string containing removal results.
         """
-        self.logger.info(
-            f"Symphony remove: removing {len(components)} components"
-        )
-        
+        self.logger.info(f"Symphony remove: removing {len(components)} components")
+
         result = SummarySpec(target_count=1, is_removal=True)
         target_result = TargetResultSpec()
         successes = 0
@@ -459,40 +408,31 @@ class MutoSymphonyProvider(BaseNode, SymphonyProvider):
 
         for component in components:
             component_name = component.name or "unnamed-component"
-            self.logger.info(
-                f"Removing component: {component_name} of type {component.type}"
-            )
+            self.logger.info(f"Removing component: {component_name} of type {component.type}")
 
             component_result = ComponentResultSpec()
 
-            stack_payload, decode_error = self._extract_stack_payload(
-                component,
-                allow_registry_lookup=True
-            )
+            stack_payload, decode_error = self._extract_stack_payload(component, allow_registry_lookup=True)
 
             if decode_error:
                 failures += 1
                 component_result.status = State.DELETE_FAILED
                 component_result.message = decode_error
                 target_result.component_results[component_name] = component_result
-                self.logger.error(
-                    f"Component {component_name} payload error: {decode_error}"
-                )
+                self.logger.error(f"Component {component_name} payload error: {decode_error}")
                 continue
 
             publish_method = self._resolve_component_method(component, default="kill")
             published = self._publish_stack_action(
                 method=publish_method,
                 payload=stack_payload,
-                context=component.metadata.get("context", "") if component.metadata else ""
+                context=component.metadata.get("context", "") if component.metadata else "",
             )
 
             if not published:
                 failures += 1
                 component_result.status = State.DELETE_FAILED
-                component_result.message = (
-                    f"Failed to publish remove action for component {component_name}"
-                )
+                component_result.message = f"Failed to publish remove action for component {component_name}"
                 target_result.component_results[component_name] = component_result
                 self.logger.error(component_result.message)
                 # Do NOT modify _component_registry on failed publish
@@ -501,9 +441,7 @@ class MutoSymphonyProvider(BaseNode, SymphonyProvider):
             # Only update registry after confirmed successful publish
             successes += 1
             component_result.status = State.DELETED
-            component_result.message = (
-                f"Remove action published for component {component_name}"
-            )
+            component_result.message = f"Remove action published for component {component_name}"
             target_result.component_results[component_name] = component_result
 
             # Registry mutation: only performed after successful stack action publish
@@ -511,9 +449,7 @@ class MutoSymphonyProvider(BaseNode, SymphonyProvider):
 
         target_result.status = "OK" if failures == 0 else "FAILED"
         if failures:
-            target_result.message = (
-                f"{failures} component(s) failed during removal"
-            )
+            target_result.message = f"{failures} component(s) failed during removal"
 
         result.success_count = successes
         if successes:
@@ -526,47 +462,37 @@ class MutoSymphonyProvider(BaseNode, SymphonyProvider):
 
         return json.dumps(result.to_dict(), indent=2)
 
-    def get(
-        self, 
-        metadata: Dict[str, Any], 
-        components: List[ComponentSpec]
-    ) -> Any:
+    def get(self, metadata: dict[str, Any], components: list[ComponentSpec]) -> Any:
         """
         Get current component states.
-        
+
         Called by MQTTBroker when Symphony requests component info.
         This method returns the current state of deployed components.
-        
+
         Args:
             metadata: Request metadata.
             components: List of components to query.
-            
+
         Returns:
             List of component data dictionaries.
         """
-        self.logger.info(
-            f"Symphony get: retrieving state for {len(components)} components"
-        )
-        
+        self.logger.info(f"Symphony get: retrieving state for {len(components)} components")
+
         # Determine which components to include in the response
         if not components:
             component_names = list(self._component_registry.keys())
-            self.logger.info(
-                "No specific components requested, returning tracked components"
-            )
+            self.logger.info("No specific components requested, returning tracked components")
         else:
             component_names = [comp.name for comp in components if comp.name]
 
         # Read-only operation: get() does NOT modify _component_registry
         # Only report components that are successfully found in the registry
-        reported_components: List[Dict[str, Any]] = []
+        reported_components: list[dict[str, Any]] = []
         for component_name in component_names:
             state_entry = self._component_registry.get(component_name)
             if not state_entry:
                 # Component not found - do not create or persist any state
-                self.logger.debug(
-                    f"Component {component_name} not found in registry; skipping"
-                )
+                self.logger.debug(f"Component {component_name} not found in registry; skipping")
                 continue
 
             # Successfully retrieved component state from registry
@@ -579,33 +505,21 @@ class MutoSymphonyProvider(BaseNode, SymphonyProvider):
                 reported_components.append(component_info)
             except (KeyError, TypeError) as exc:
                 # Failed to retrieve component state - do not persist stale data
-                self.logger.warning(
-                    f"Failed to retrieve state for component {component_name}: {exc}"
-                )
+                self.logger.warning(f"Failed to retrieve state for component {component_name}: {exc}")
                 continue
 
         return json.dumps(reported_components, indent=2)
 
-    def _resolve_component_method(
-        self,
-        component: ComponentSpec,
-        default: str
-    ) -> str:
+    def _resolve_component_method(self, component: ComponentSpec, default: str) -> str:
         """Determine the composer method for the provided component."""
         props = component.properties or {}
         method = props.get("method") or props.get("action")
 
         if not method and component.parameters:
-            method = (
-                component.parameters.get("method")
-                or component.parameters.get("action")
-            )
+            method = component.parameters.get("method") or component.parameters.get("action")
 
         if not method and component.metadata:
-            method = (
-                component.metadata.get("method")
-                or component.metadata.get("action")
-            )
+            method = component.metadata.get("method") or component.metadata.get("action")
 
         resolved = (method or default).lower()
 
@@ -616,10 +530,8 @@ class MutoSymphonyProvider(BaseNode, SymphonyProvider):
         return resolved
 
     def _extract_stack_payload(
-        self,
-        component: ComponentSpec,
-        allow_registry_lookup: bool = False
-    ) -> Tuple[Optional[Any], Optional[str]]:
+        self, component: ComponentSpec, allow_registry_lookup: bool = False
+    ) -> tuple[Any | None, str | None]:
         """Extract a JSON payload representing the stack for a component."""
         try:
             props = component.properties or {}
@@ -639,7 +551,7 @@ class MutoSymphonyProvider(BaseNode, SymphonyProvider):
             elif isinstance(data, str):
                 decoded_bytes = self._attempt_base64_decode(data)
                 if decoded_bytes is None:
-                    decoded_bytes = data.encode('utf-8')
+                    decoded_bytes = data.encode("utf-8")
             else:
                 if allow_registry_lookup:
                     registry_entry = self._component_registry.get(component.name or "")
@@ -647,7 +559,7 @@ class MutoSymphonyProvider(BaseNode, SymphonyProvider):
                         return registry_entry.get("payload"), None
                 return None, "Unsupported payload format"
 
-            payload_str = decoded_bytes.decode('utf-8')
+            payload_str = decoded_bytes.decode("utf-8")
             return json.loads(payload_str), None
 
         except json.JSONDecodeError as exc:
@@ -655,19 +567,14 @@ class MutoSymphonyProvider(BaseNode, SymphonyProvider):
         except Exception as exc:
             return None, f"Unexpected error reading stack payload: {exc}"
 
-    def _attempt_base64_decode(self, data: str) -> Optional[bytes]:
+    def _attempt_base64_decode(self, data: str) -> bytes | None:
         """Try to base64 decode a string, returning None if decoding fails."""
         try:
             return base64.b64decode(data)
         except Exception:
             return None
 
-    def _publish_stack_action(
-        self,
-        method: str,
-        payload: Any,
-        context: str = ""
-    ) -> bool:
+    def _publish_stack_action(self, method: str, payload: Any, context: str = "") -> bool:
         """Publish a stack action to the composer for execution."""
         if not self.stack_publisher:
             self.logger.error("Stack publisher is not initialized")
@@ -682,69 +589,58 @@ class MutoSymphonyProvider(BaseNode, SymphonyProvider):
             msg_action.payload = payload_str
 
             self.stack_publisher.publish(msg_action)
-            self.logger.debug(
-                f"Published stack action '{method}' with payload length {len(payload_str)}"
-            )
+            self.logger.debug(f"Published stack action '{method}' with payload length {len(payload_str)}")
             return True
         except Exception as exc:
             self.logger.error(f"Failed to publish stack action '{method}': {exc}")
             return False
 
-    def needs_update(
-        self, 
-        metadata: Dict[str, Any], 
-        pack: ComparisonPack
-    ) -> bool:
+    def needs_update(self, metadata: dict[str, Any], pack: ComparisonPack) -> bool:
         """
         Check if components need updates by comparing desired vs current state.
-        
+
         Args:
             metadata: Request metadata.
             pack: Comparison pack containing desired and current components.
-            
+
         Returns:
             True if updates are needed, False otherwise.
         """
-        self.logger.info(
-            f"Checking update need for {len(pack.desired)} desired vs "
-            f"{len(pack.current)} current"
-        )
-        
+        self.logger.info(f"Checking update need for {len(pack.desired)} desired vs {len(pack.current)} current")
+
         # Create lookup for current components
         current_by_name = {comp.name: comp for comp in pack.current}
-        
+
         for desired in pack.desired:
             current = current_by_name.get(desired.name)
-            
+
             if not current:
                 # Component doesn't exist, needs deployment
-                self.logger.info(
-                    f"Component {desired.name} not found - needs deployment"
-                )
+                self.logger.info(f"Component {desired.name} not found - needs deployment")
                 return True
-            
+
             # Check if component properties have changed
             if self._component_changed(desired, current):
                 self.logger.info(f"Component {desired.name} has changes - needs update")
                 return True
-        
+
         # No updates needed
         self.logger.info("No components need updates")
         return False
 
-    def needs_remove(self, metadata: Dict[str, Any], pack: ComparisonPack) -> bool:
+    def needs_remove(self, metadata: dict[str, Any], pack: ComparisonPack) -> bool:
         """Check if components need removal by comparing desired vs current state."""
         self.logger.info(f"Checking removal need for {len(pack.desired)} desired vs {len(pack.current)} current")
-        
+
         # Create lookup for desired components
         desired_by_name = {comp.name: comp for comp in pack.desired}
-        
+
         for current in pack.current:
             if current.name not in desired_by_name:
                 # Current component not in desired state, needs removal
                 self.logger.info(f"Component {current.name} not desired - needs removal")
                 return True
-        
+
         # No removals needed
         self.logger.info("No components need removal")
         return False
@@ -752,46 +648,43 @@ class MutoSymphonyProvider(BaseNode, SymphonyProvider):
     def _component_changed(self, desired: ComponentSpec, current: ComponentSpec) -> bool:
         """Check if a component has changed between desired and current state."""
         # Compare key properties that would require updates
-        
+
         # Type change
         if desired.type != current.type:
             return True
-        
+
         # Properties change
         if desired.properties != current.properties:
             return True
-        
+
         # Parameters change
         if desired.parameters != current.parameters:
             return True
-        
+
         # Constraints change
         if desired.constraints != current.constraints:
             return True
-        
+
         # Dependencies change
-        if desired.dependencies != current.dependencies:
-            return True
-        
-        return False
+        return desired.dependencies != current.dependencies
 
     def start(self) -> bool:
         """Start the Symphony provider."""
         if not self._config.symphony.enabled:
             self.logger.info("Symphony provider not enabled in configuration")
             return False
-        
+
         if self._running:
             self.logger.warning("Symphony provider already running")
             return True
-        
+
         try:
             # The MQTT broker is initialized in __init__ if enabled
             # No additional startup steps needed
             self._running = True
             self.logger.info("Muto Symphony Provider started successfully")
             return True
-            
+
         except Exception as e:
             self.logger.error(f"Failed to start Symphony provider: {e}")
             return False
@@ -800,19 +693,19 @@ class MutoSymphonyProvider(BaseNode, SymphonyProvider):
         """Stop the Symphony provider."""
         if not self._running:
             return  # Already cleaned up
-            
+
         self._running = False
         self.logger.info("Stopping Muto Symphony Provider")
-        
+
         # Set shutdown event first to signal other threads
         self._shutdown_event.set()
-        
+
         try:
             # Unregister from Symphony
             self.unregister_target()
         except Exception as e:
             self.logger.error(f"Error during Symphony unregistration: {e}")
-        
+
         try:
             # Stop MQTT broker
             if self._mqtt_broker:
@@ -820,7 +713,7 @@ class MutoSymphonyProvider(BaseNode, SymphonyProvider):
                 self._mqtt_broker = None
         except Exception as e:
             self.logger.error(f"Error stopping MQTT broker: {e}")
-        
+
         try:
             # Close API client
             if self._api_client:
@@ -828,18 +721,18 @@ class MutoSymphonyProvider(BaseNode, SymphonyProvider):
                 self._api_client = None
         except Exception as e:
             self.logger.error(f"Error closing API client: {e}")
-        
+
         try:
             # Destroy ROS2 services and publishers
-            if hasattr(self, '_register_service') and self._register_service:
+            if hasattr(self, "_register_service") and self._register_service:
                 self.destroy_service(self._register_service)
-            if hasattr(self, '_unregister_service') and self._unregister_service:
+            if hasattr(self, "_unregister_service") and self._unregister_service:
                 self.destroy_service(self._unregister_service)
-            if hasattr(self, 'stack_publisher') and self.stack_publisher:
+            if hasattr(self, "stack_publisher") and self.stack_publisher:
                 self.destroy_publisher(self.stack_publisher)
         except Exception as e:
             self.logger.error(f"Error destroying ROS2 resources: {e}")
-        
+
         self.logger.info("Muto Symphony Provider stopped")
 
     def is_running(self) -> bool:
@@ -857,43 +750,41 @@ class MutoSymphonyProvider(BaseNode, SymphonyProvider):
         return 0
 
 
-
-
 def main():
     """Main entry point for the Symphony Provider."""
     provider = None
     shutdown_requested = threading.Event()
-    
+
     def signal_handler(signum, frame):
         """Handle shutdown signals gracefully."""
         print(f"Received signal {signum}, initiating graceful shutdown...")
         shutdown_requested.set()
         if provider is not None:
             provider._shutdown_event.set()
-    
+
     # Register signal handlers
     signal.signal(signal.SIGINT, signal_handler)
     signal.signal(signal.SIGTERM, signal_handler)
-    
+
     try:
         rclpy.init()
         provider = MutoSymphonyProvider()
         provider.initialize()
-        
+
         provider.get_logger().info("Symphony Provider started successfully")
-        
+
         # Custom spin loop to handle shutdown gracefully
         while rclpy.ok() and not shutdown_requested.is_set():
             try:
                 rclpy.spin_once(provider, timeout_sec=1.0)
             except KeyboardInterrupt:
                 break
-        
+
     except KeyboardInterrupt:
         print("Symphony Provider interrupted by user")
     except Exception as e:
         print(f"Failed to start Symphony Provider: {e}")
-        
+
     finally:
         # Cleanup provider if it was created
         if provider is not None:
@@ -902,7 +793,7 @@ def main():
                 provider.cleanup()
             except Exception as e:
                 print(f"Error during provider cleanup: {e}")
-        
+
         # Only shutdown ROS2 if it's still initialized and we haven't already shut it down
         try:
             if rclpy.ok():
@@ -911,5 +802,6 @@ def main():
         except Exception as e:
             print(f"Error during ROS2 shutdown (this may be normal): {e}")
 
-if __name__ == '__main__':
+
+if __name__ == "__main__":
     exit(main())
