@@ -30,55 +30,62 @@ import json
 import logging
 from dataclasses import dataclass, field
 from enum import Enum
-from typing import Any, Dict, List, Optional, Union, get_args, get_origin
+from typing import Any, Optional, get_args, get_origin
 
 from symphony_sdk.types import State
 
-
 logger = logging.getLogger(__name__)
+
 
 @dataclass
 class ObjectMeta:
     """Object metadata following Kubernetes-style metadata."""
+
     namespace: str = ""
     name: str = ""
-    labels: Optional[Dict[str, str]] = None
-    annotations: Optional[Dict[str, str]] = None
+    labels: dict[str, str] | None = None
+    annotations: dict[str, str] | None = None
 
 
 @dataclass
 class TargetSelector:
     """Target selector for component binding."""
+
     name: str = ""
-    selector: Optional[Dict[str, str]] = None
+    selector: dict[str, str] | None = None
 
 
 @dataclass
 class BindingSpec:
     """Binding specification for component deployment."""
+
     role: str = ""
     provider: str = ""
-    config: Optional[Dict[str, str]] = None
+    config: dict[str, str] | None = None
+
 
 @dataclass
 class TopologySpec:
     """Topology specification for deployment."""
+
     device: str = ""
-    selector: Optional[Dict[str, str]] = None
-    bindings: Optional[List[BindingSpec]] = None
+    selector: dict[str, str] | None = None
+    bindings: list[BindingSpec] | None = None
 
 
 @dataclass
 class PipelineSpec:
     """Pipeline specification for data processing."""
+
     name: str = ""
     skill: str = ""
-    parameters: Optional[Dict[str, str]] = None
+    parameters: dict[str, str] | None = None
 
 
 @dataclass
 class VersionSpec:
     """Version specification for solutions."""
+
     solution: str = ""
     percentage: int = 100
 
@@ -86,89 +93,99 @@ class VersionSpec:
 @dataclass
 class InstanceSpec:
     """Instance specification for Symphony deployments."""
+
     name: str = ""
-    parameters: Optional[Dict[str, str]] = None
+    parameters: dict[str, str] | None = None
     solution: str = ""
-    target: Optional[TargetSelector] = None
-    topologies: Optional[List[TopologySpec]] = None
-    pipelines: Optional[List[PipelineSpec]] = None
+    target: TargetSelector | None = None
+    topologies: list[TopologySpec] | None = None
+    pipelines: list[PipelineSpec] | None = None
     scope: str = ""
     display_name: str = ""
-    metadata: Optional[Dict[str, str]] = None
-    versions: Optional[List[VersionSpec]] = None
-    arguments: Optional[Dict[str, Dict[str, str]]] = None
+    metadata: dict[str, str] | None = None
+    versions: list[VersionSpec] | None = None
+    arguments: dict[str, dict[str, str]] | None = None
     opt_out_reconciliation: bool = False
 
 
 @dataclass
 class FilterSpec:
     """Filter specification for routing."""
+
     direction: str = ""
-    parameters: Optional[Dict[str, str]] = None
+    parameters: dict[str, str] | None = None
     type: str = ""
+
 
 @dataclass
 class RouteSpec:
     route: str = ""
-    properties: Dict[str, str] = None
-    filters: List[FilterSpec] = None
+    properties: dict[str, str] = None
+    filters: list[FilterSpec] = None
     type: str = ""
+
 
 @dataclass
 class ComponentSpec:
-    name: str = ""   
+    name: str = ""
     type: str = ""
-    routes: List[RouteSpec] = None
+    routes: list[RouteSpec] = None
     constraints: str = ""
-    properties: Dict[str, str] = None
-    depedencies: List[str] = None
-    skills: List[str] = None
-    metadata: Dict[str, str] = None
-    parameters: Dict[str, str] = None
-    
+    properties: dict[str, str] = None
+    depedencies: list[str] = None
+    skills: list[str] = None
+    metadata: dict[str, str] = None
+    parameters: dict[str, str] = None
+
 
 @dataclass
-class SolutionSpec:    
-    components: List[ComponentSpec] = None    
-    scope: str = ""    
+class SolutionSpec:
+    components: list[ComponentSpec] = None
+    scope: str = ""
     displayName: str = ""
-    metadata: Dict[str,str] = None
-    
+    metadata: dict[str, str] = None
+
+
 @dataclass
 class SolutionState:
     metadata: ObjectMeta = None
     spec: SolutionSpec = None
 
+
 @dataclass
 class TargetSpec:
-    properties: Dict[str, str] = None
-    components: List[ComponentSpec] = None
+    properties: dict[str, str] = None
+    components: list[ComponentSpec] = None
     constraints: str = ""
-    topologies: List[TopologySpec] = None
+    topologies: list[TopologySpec] = None
     scope: str = ""
     displayName: str = ""
-    metadata: Dict[str, str] = None
+    metadata: dict[str, str] = None
     forceRedeploy: bool = False
+
 
 @dataclass
 class ComponentError:
     code: str = ""
     message: str = ""
-    target: str = ""    
+    target: str = ""
+
 
 @dataclass
 class TargetError:
     code: str = ""
     message: str = ""
     target: str = ""
-    details: Dict[str, ComponentError] = None
+    details: dict[str, ComponentError] = None
+
 
 @dataclass
 class ErrorType:
     code: str = ""
     message: str = ""
     target: str = ""
-    details: Dict[str, TargetError] = None
+    details: dict[str, TargetError] = None
+
 
 @dataclass
 class ProvisioningStatus:
@@ -177,13 +194,15 @@ class ProvisioningStatus:
     failureCause: str = ""
     logErrors: bool = False
     error: ErrorType = None
-    output: Dict[str, str] = None
+    output: dict[str, str] = None
+
 
 @dataclass
 class TargetStatus:
-    properties: Dict[str, str] = None
+    properties: dict[str, str] = None
     provisioningStatus: ProvisioningStatus = None
     lastModififed: str = ""
+
 
 @dataclass
 class TargetState:
@@ -191,54 +210,62 @@ class TargetState:
     spec: TargetSpec = None
     status: TargetStatus = None
 
+
 @dataclass
 class DeviceSpec:
-    properties: Dict[str, str] = None
-    bindings: List[BindingSpec] = None
+    properties: dict[str, str] = None
+    bindings: list[BindingSpec] = None
     displayName: str = ""
-    
+
+
 @dataclass
 class DeploymentSpec:
     solutionName: str = ""
     solution: SolutionState = None
     instance: InstanceSpec = None
-    targets: Dict[str, TargetState] = None
-    devices: List[DeviceSpec] = None
-    assignments: Dict[str, str] = None
+    targets: dict[str, TargetState] = None
+    devices: list[DeviceSpec] = None
+    assignments: dict[str, str] = None
     componentStartIndex: int = -1
-    componentEndIndex: int= -1
+    componentEndIndex: int = -1
     activeTarget: str = ""
 
-    def get_components_slice(self) -> List[ComponentSpec]:
-        if self.solution != None:
-            if self.componentStartIndex >= 0 and self.componentEndIndex >= 0 and self.componentEndIndex > self.componentStartIndex:
-                return self.solution.spec.components[self.componentStartIndex: self.componentEndIndex]
+    def get_components_slice(self) -> list[ComponentSpec]:
+        if self.solution is not None:
+            if (
+                self.componentStartIndex >= 0
+                and self.componentEndIndex >= 0
+                and self.componentEndIndex > self.componentStartIndex
+            ):
+                return self.solution.spec.components[self.componentStartIndex : self.componentEndIndex]
             return self.solution.spec.components
         return []
 
+
 @dataclass
 class ComparisonPack:
-    desired: List[ComponentSpec]
-    current: List[ComponentSpec]
+    desired: list[ComponentSpec]
+    current: list[ComponentSpec]
 
 
 @dataclass
 class COABodyMixin:
     """
     Common functionality for COA request and response body handling.
-    
+
     Provides content-type aware body encoding/decoding with support for:
     - "application/json": JSON objects
     - "text/plain": Plain text strings
     - "application/octet-stream": Binary data
     """
+
     content_type: str = "application/json"
     body: str = ""  # Base64 encoded data (type determined by content_type)
-    
-    def set_body(self, data: Any, content_type: Optional[str] = None) -> None:
+
+    def set_body(self, data: Any, content_type: str | None = None) -> None:
         """
         Set body data with content type detection or explicit content type.
-        
+
         Args:
             data: Data to set (JSON object, string, or bytes)
             content_type: Optional explicit content type override
@@ -246,27 +273,27 @@ class COABodyMixin:
         # Set content type if provided
         if content_type:
             self.content_type = content_type
-        
+
         if self.content_type == "application/json":
             # Handle JSON data
             if isinstance(data, (str, bytes)):
                 # If it's a string/bytes, try to parse as JSON first
                 if isinstance(data, bytes):
-                    data = data.decode('utf-8')
+                    data = data.decode("utf-8")
                 json.loads(data)  # Validate JSON
-                self.body = base64.b64encode(data.encode('utf-8')).decode('utf-8')
+                self.body = base64.b64encode(data.encode("utf-8")).decode("utf-8")
             else:
                 # Serialize object to JSON
                 json_str = json.dumps(data, ensure_ascii=False)
-                self.body = base64.b64encode(json_str.encode('utf-8')).decode('utf-8')
-        
+                self.body = base64.b64encode(json_str.encode("utf-8")).decode("utf-8")
+
         elif self.content_type == "text/plain":
             # Handle plain text - no encoding necessary
             if isinstance(data, bytes):
-                self.body = data.decode('utf-8')
+                self.body = data.decode("utf-8")
             else:
                 self.body = str(data)
-        
+
         elif self.content_type == "application/octet-stream":
             # Handle binary data
             if isinstance(data, str):
@@ -274,72 +301,70 @@ class COABodyMixin:
                 self.body = data
             elif isinstance(data, bytes):
                 # Encode bytes to base64
-                self.body = base64.b64encode(data).decode('utf-8')
+                self.body = base64.b64encode(data).decode("utf-8")
             else:
                 raise ValueError(f"Binary content type requires bytes or base64 string, got {type(data)}")
-        
+
         else:
             # Default fallback - treat as text
-            if isinstance(data, bytes):
-                text_str = data.decode('utf-8')
-            else:
-                text_str = str(data)
-            self.body = base64.b64encode(text_str.encode('utf-8')).decode('utf-8')
-    
+            text_str = data.decode("utf-8") if isinstance(data, bytes) else str(data)
+            self.body = base64.b64encode(text_str.encode("utf-8")).decode("utf-8")
+
     def get_body(self) -> Any:
         """
         Get body data decoded according to content type.
-        
+
         Returns:
             Decoded body data (JSON object, string, or bytes depending on content_type)
         """
         if not self.body:
             return None
-        
+
         if self.content_type == "application/json":
             # Return parsed JSON object
             try:
-                json_str = base64.b64decode(self.body).decode('utf-8')
+                json_str = base64.b64decode(self.body).decode("utf-8")
                 return json.loads(json_str)
             except (ValueError, json.JSONDecodeError) as e:
-                raise ValueError(f"Invalid JSON in body: {e}")
-        
+                raise ValueError(f"Invalid JSON in body: {e}") from e
+
         elif self.content_type == "text/plain":
             # Return text string directly - no decoding necessary
             return self.body
-        
+
         elif self.content_type == "application/octet-stream":
             # Return raw bytes
             return base64.b64decode(self.body)
-        
+
         else:
             # Default fallback - return as string
-            return base64.b64decode(self.body).decode('utf-8')
+            return base64.b64decode(self.body).decode("utf-8")
 
 
-@dataclass  
+@dataclass
 class COARequest(COABodyMixin):
     """
     COA Request structure based on Symphony COA API.
-    
+
     This corresponds to the Go struct COARequest from Symphony codebase.
     The body field contains base64 encoded data, with the content type determined by content_type:
     - "application/json": Base64 encoded UTF-8 string of JSON object
-    - "text/plain": Base64 encoded UTF-8 string of plain text  
+    - "text/plain": Base64 encoded UTF-8 string of plain text
     - "application/octet-stream": Base64 encoded binary data
     """
+
     method: str = "GET"
     route: str = ""
-    metadata: Optional[Dict[str, str]] = field(default_factory=dict)
-    parameters: Optional[Dict[str, str]] = field(default_factory=dict)
-    
-    def to_json_dict(self) -> Dict[str, Any]:
+    metadata: dict[str, str] | None = field(default_factory=dict)
+    parameters: dict[str, str] | None = field(default_factory=dict)
+
+    def to_json_dict(self) -> dict[str, Any]:
         """Convert to JSON-serializable dictionary."""
         result = {
             "method": self.method,
             "route": self.route,
             "content-type": self.content_type,
-            "body": self.body
+            "body": self.body,
         }
         if self.metadata:
             result["metadata"] = self.metadata
@@ -352,41 +377,42 @@ class COARequest(COABodyMixin):
 class COAResponse(COABodyMixin):
     """
     COA Response structure based on Symphony COA API.
-    
+
     This corresponds to the Go struct COAResponse from Symphony codebase.
     The body field contains base64 encoded data, with the content type determined by content_type:
     - "application/json": Base64 encoded UTF-8 string of JSON object
-    - "text/plain": Base64 encoded UTF-8 string of plain text  
+    - "text/plain": Base64 encoded UTF-8 string of plain text
     - "application/octet-stream": Base64 encoded binary data
     """
+
     state: State = State.OK
-    metadata: Optional[Dict[str, str]] = field(default_factory=dict)
-    redirect_uri: Optional[str] = None
-    
-    def to_json_dict(self) -> Dict[str, Any]:
+    metadata: dict[str, str] | None = field(default_factory=dict)
+    redirect_uri: str | None = None
+
+    def to_json_dict(self) -> dict[str, Any]:
         """Convert to JSON-serializable dictionary."""
-        result = {
-            "content-type": self.content_type,
-            "body": self.body,
-            "state": self.state.value
-        }
+        result = {"content-type": self.content_type, "body": self.body, "state": self.state.value}
         if self.metadata:
             result["metadata"] = self.metadata
         if self.redirect_uri:
             result["redirectUri"] = self.redirect_uri
         return result
-    
+
     @classmethod
-    def success(cls, data: Any = None, content_type: str = "application/json") -> 'COAResponse':
+    def success(cls, data: Any = None, content_type: str = "application/json") -> "COAResponse":
         """Create a success response."""
         response = cls(content_type=content_type, state=State.OK)
         if data is not None:
             response.set_body(data, content_type)
         return response
-    
+
     @classmethod
-    def error(cls, message: str, state: State = State.INTERNAL_ERROR, 
-              content_type: str = "application/json") -> 'COAResponse':
+    def error(
+        cls,
+        message: str,
+        state: State = State.INTERNAL_ERROR,
+        content_type: str = "application/json",
+    ) -> "COAResponse":
         """Create an error response."""
         response = cls(content_type=content_type, state=state)
         if content_type == "application/json":
@@ -396,25 +422,25 @@ class COAResponse(COABodyMixin):
         else:
             response.set_body({"error": message}, "application/json")
         return response
-    
+
     @classmethod
-    def not_found(cls, message: str = "Resource not found") -> 'COAResponse':
+    def not_found(cls, message: str = "Resource not found") -> "COAResponse":
         """Create a not found response."""
         return cls.error(message, State.NOT_FOUND)
-    
+
     @classmethod
-    def bad_request(cls, message: str = "Bad request") -> 'COAResponse':
+    def bad_request(cls, message: str = "Bad request") -> "COAResponse":
         """Create a bad request response."""
         return cls.error(message, State.BAD_REQUEST)
 
 
 # Utility functions for COA data conversion
-def to_dict(obj: Any) -> Dict[str, Any]:
+def to_dict(obj: Any) -> dict[str, Any]:
     """Convert dataclass object to dictionary."""
     if obj is None:
         return {}
-    
-    if hasattr(obj, '__dict__'):
+
+    if hasattr(obj, "__dict__"):
         result = {}
         for key, value in obj.__dict__.items():
             if value is not None:
@@ -422,39 +448,39 @@ def to_dict(obj: Any) -> Dict[str, Any]:
                     result[key] = [to_dict(item) for item in value]
                 elif isinstance(value, dict):
                     result[key] = {k: to_dict(v) for k, v in value.items()}
-                elif hasattr(value, '__dict__'):
+                elif hasattr(value, "__dict__"):
                     result[key] = to_dict(value)
                 elif isinstance(value, Enum):
                     result[key] = value.value
                 else:
                     result[key] = value
         return result
-    
+
     return obj
 
 
-def from_dict(data: Dict[str, Any], cls: type) -> Any:
+def from_dict(data: dict[str, Any], cls: type) -> Any:
     """Convert dictionary to dataclass object."""
     if not data:
         return cls()
-    
+
     try:
         # Handle enums
         if isinstance(cls, type) and issubclass(cls, Enum):
             return cls(data)
-        
+
         # Handle basic types
         if cls in (str, int, float, bool):
             return cls(data)
-        
+
         # Handle dataclass
-        if hasattr(cls, '__dataclass_fields__'):
+        if hasattr(cls, "__dataclass_fields__"):
             kwargs = {}
             for field_name, field_info in cls.__dataclass_fields__.items():
                 if field_name in data:
                     field_type = field_info.type
                     field_value = data[field_name]
-                    
+
                     # Handle Optional types
                     if get_origin(field_type) is Optional and type(None) in get_args(field_type):
                         if field_value is None:
@@ -462,43 +488,43 @@ def from_dict(data: Dict[str, Any], cls: type) -> Any:
                         else:
                             inner_type = field_type.__args__[0]
                             kwargs[field_name] = from_dict(field_value, inner_type)
-                    
+
                     # Handle List types
-                    elif hasattr(field_type, '__origin__') and field_type.__origin__ is list:
+                    elif hasattr(field_type, "__origin__") and field_type.__origin__ is list:
                         if field_value and isinstance(field_value, list):
                             inner_type = field_type.__args__[0]
                             kwargs[field_name] = [from_dict(item, inner_type) for item in field_value]
                         else:
                             kwargs[field_name] = field_value or []
-                    
+
                     # Handle Dict types
-                    elif hasattr(field_type, '__origin__') and field_type.__origin__ is dict:
+                    elif hasattr(field_type, "__origin__") and field_type.__origin__ is dict:
                         kwargs[field_name] = field_value or {}
-                    
+
                     # Handle nested dataclasses
-                    elif hasattr(field_type, '__dataclass_fields__'):
+                    elif hasattr(field_type, "__dataclass_fields__"):
                         kwargs[field_name] = from_dict(field_value, field_type)
-                    
+
                     # Handle enums
                     elif isinstance(field_type, type) and issubclass(field_type, Enum):
                         kwargs[field_name] = field_type(field_value)
-                    
+
                     else:
                         kwargs[field_name] = field_value
-            
+
             return cls(**kwargs)
-    
-    except Exception as e:
+
+    except Exception:
         # Fallback: return default instance
         return cls()
 
 
-def serialize_components(components: List[ComponentSpec]) -> str:
+def serialize_components(components: list[ComponentSpec]) -> str:
     """Serialize components to JSON string."""
     return json.dumps([to_dict(comp) for comp in components], indent=2)
 
 
-def deserialize_components(json_str: str) -> List[ComponentSpec]:
+def deserialize_components(json_str: str) -> list[ComponentSpec]:
     """Deserialize JSON string to components list."""
     try:
         data = json.loads(json_str)
@@ -507,7 +533,7 @@ def deserialize_components(json_str: str) -> List[ComponentSpec]:
         return []
 
 
-def deserialize_solution(json_str: str) -> List[SolutionState]:
+def deserialize_solution(json_str: str) -> list[SolutionState]:
     """Deserialize JSON string to components list."""
     try:
         data = json.loads(json_str)
@@ -516,8 +542,8 @@ def deserialize_solution(json_str: str) -> List[SolutionState]:
         return []
 
 
-# Desrialize a DeploymentSpec object  from Json String 
-def deserialize_deployment(json_str: str) -> List[DeploymentSpec]:
+# Desrialize a DeploymentSpec object  from Json String
+def deserialize_deployment(json_str: str) -> list[DeploymentSpec]:
     """Deserialize JSON string to DeploymentSpec list."""
     try:
         data = json.loads(json_str)
@@ -537,12 +563,12 @@ def deserialize_coa_request(json_str: str) -> COARequest:
     try:
         data = json.loads(json_str)
         request = COARequest()
-        
+
         # Map JSON fields to dataclass fields
         if "method" in data:
             request.method = data["method"]
         if "route" in data:
-            request.route = data["route"]  
+            request.route = data["route"]
         if "content-type" in data:
             request.content_type = data["content-type"]
         if "body" in data:
@@ -557,7 +583,7 @@ def deserialize_coa_request(json_str: str) -> COARequest:
             request.metadata = data["metadata"]
         if "parameters" in data:
             request.parameters = data["parameters"]
-            
+
         return request
     except Exception as e:
         logger.error("Error deserializing COA request: %s", e)
@@ -574,7 +600,7 @@ def deserialize_coa_response(json_str: str) -> COAResponse:
     try:
         data = json.loads(json_str)
         response = COAResponse()
-        
+
         # Map JSON fields to dataclass fields
         if "content-type" in data:
             response.content_type = data["content-type"]
@@ -595,7 +621,7 @@ def deserialize_coa_response(json_str: str) -> COAResponse:
             response.metadata = data["metadata"]
         if "redirectUri" in data:
             response.redirect_uri = data["redirectUri"]
-            
+
         return response
     except Exception as e:
         logger.error("Error deserializing COA response: %s", e)
@@ -603,39 +629,39 @@ def deserialize_coa_response(json_str: str) -> COAResponse:
 
 
 __all__ = [
-    'ObjectMeta',
-    'TargetSelector',
-    'BindingSpec',
-    'TopologySpec',
-    'PipelineSpec',
-    'VersionSpec',
-    'InstanceSpec',
-    'FilterSpec',
-    'RouteSpec',
-    'ComponentSpec',
-    'SolutionSpec',
-    'SolutionState',
-    'TargetSpec',
-    'ComponentError',
-    'TargetError',
-    'ErrorType',
-    'ProvisioningStatus',
-    'TargetStatus',
-    'TargetState',
-    'DeviceSpec',
-    'DeploymentSpec',
-    'ComparisonPack',
-    'COABodyMixin',
-    'COARequest',
-    'COAResponse',
-    'to_dict',
-    'from_dict',
-    'serialize_components',
-    'deserialize_components',
-    'deserialize_solution',
-    'deserialize_deployment',
-    'serialize_coa_request',
-    'deserialize_coa_request',
-    'serialize_coa_response',
-    'deserialize_coa_response',
+    "ObjectMeta",
+    "TargetSelector",
+    "BindingSpec",
+    "TopologySpec",
+    "PipelineSpec",
+    "VersionSpec",
+    "InstanceSpec",
+    "FilterSpec",
+    "RouteSpec",
+    "ComponentSpec",
+    "SolutionSpec",
+    "SolutionState",
+    "TargetSpec",
+    "ComponentError",
+    "TargetError",
+    "ErrorType",
+    "ProvisioningStatus",
+    "TargetStatus",
+    "TargetState",
+    "DeviceSpec",
+    "DeploymentSpec",
+    "ComparisonPack",
+    "COABodyMixin",
+    "COARequest",
+    "COAResponse",
+    "to_dict",
+    "from_dict",
+    "serialize_components",
+    "deserialize_components",
+    "deserialize_solution",
+    "deserialize_deployment",
+    "serialize_coa_request",
+    "deserialize_coa_request",
+    "serialize_coa_response",
+    "deserialize_coa_response",
 ]
