@@ -1,4 +1,3 @@
-# -*- coding: utf-8 -*-
 #
 # Copyright (c) 2019-2022, Martin Günther (DFKI GmbH) and others
 # Copyright (c) 2013-2016, Brandon Alexander
@@ -17,7 +16,7 @@
 
 import json
 
-from agent.ros.msg_converter import message_converter
+from muto_agent.ros.msg_converter import message_converter
 
 
 def convert_json_to_ros_message(message_type, json_message, strict_mode=True):
