@@ -1,28 +1,21 @@
 #
-#  Copyright (c) 2023 Composiv.ai
+# Copyright (c) 2023 Composiv.ai
 #
-# All rights reserved. This program and the accompanying materials
-# are made available under the terms of the Eclipse Public License v2.0
-# and Eclipse Distribution License v1.0 which accompany this distribution.
+# This program and the accompanying materials are made available under the
+# terms of the Eclipse Public License 2.0 which is available at
+# http://www.eclipse.org/legal/epl-2.0.
 #
-# Licensed under the  Eclipse Public License v2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-# The Eclipse Public License is available at
-#    http://www.eclipse.org/legal/epl-v20.html
-# and the Eclipse Distribution License is available at
-#   http://www.eclipse.org/org/documents/edl-v10.php.
+# SPDX-License-Identifier: EPL-2.0
 #
 # Contributors:
-#    Composiv.ai - initial API and implementation
+#   Composiv.ai - initial API and implementation
 #
-#
-from muto_msgs.srv import CommandPlugin
-
 import json
 
+from muto_msgs.srv import CommandPlugin
 
-class ParamCommands():
+
+class ParamCommands:
     """
     Class for managing rosparam commands.
 
@@ -91,7 +84,7 @@ class ParamCommands():
         try:
             value = self.node.get_parameter(requested_param).value
             result = {"name": requested_param, "value": value}
-        except:
+        except Exception:
             self.node.get_logger().error(f"{requested_param} not found.")
 
         response.output = self.node.construct_command_output_message(result)
