@@ -1,17 +1,16 @@
 """Public Symphony SDK interface."""
 
+from . import models as _models
+from . import summary as _summary
+from . import types as _types
 from .api_client import SymphonyAPI, SymphonyAPIError
 from .models import *  # noqa: F401,F403
 from .summary import *  # noqa: F401,F403
 from .types import *  # noqa: F401,F403
 
-from . import models as _models
-from . import summary as _summary
-from . import types as _types
-
 __all__ = [
-    'SymphonyAPI',
-    'SymphonyAPIError',
+    "SymphonyAPI",
+    "SymphonyAPIError",
     *_models.__all__,
     *_summary.__all__,
     *_types.__all__,
