@@ -1,29 +1,24 @@
 #
-#  Copyright (c) 2023 Composiv.ai
+# Copyright (c) 2023 Composiv.ai
 #
-# All rights reserved. This program and the accompanying materials
-# are made available under the terms of the Eclipse Public License v2.0
-# and Eclipse Distribution License v1.0 which accompany this distribution.
+# This program and the accompanying materials are made available under the
+# terms of the Eclipse Public License 2.0 which is available at
+# http://www.eclipse.org/legal/epl-2.0.
 #
-# Licensed under the  Eclipse Public License v2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-# The Eclipse Public License is available at
-#    http://www.eclipse.org/legal/epl-v20.html
-# and the Eclipse Distribution License is available at
-#   http://www.eclipse.org/org/documents/edl-v10.php.
+# SPDX-License-Identifier: EPL-2.0
 #
 # Contributors:
-#    Composiv.ai - initial API and implementation
-#
+#   Composiv.ai - initial API and implementation
 #
 
-import unittest
+import contextlib
 import json
+import unittest
 
 import rclpy
-import agent.muto_agent
 from muto_msgs.msg import Gateway, MutoAction, MutoActionMeta
+
+import muto_agent.muto_agent
 
 
 class TestAgentNode(unittest.TestCase):
@@ -36,7 +31,7 @@ class TestAgentNode(unittest.TestCase):
         rclpy.shutdown()
 
     def setUp(self):
-        self.node = agent.muto_agent.MutoAgent()
+        self.node = muto_agent.muto_agent.MutoAgent()
         # Initialize the node to setup configuration and parameters
         try:
             self.node.initialize()
@@ -45,14 +40,12 @@ class TestAgentNode(unittest.TestCase):
             pass
 
     def tearDown(self):
-        try:
+        with contextlib.suppress(BaseException):
             self.node.cleanup()
-        except:
-            pass
         self.node.destroy_node()
 
     def test_mqtt_node_create(self):
-        assert self.node != None, "Node couldn't be created."
+        assert self.node is not None, "Node couldn't be created."
 
     def test_gateway_msg_callback_ping(self):
         meta_msg = MutoActionMeta()
@@ -67,10 +60,10 @@ class TestAgentNode(unittest.TestCase):
                 "headers": {
                     "content-type": "application/json",
                     "reply-to": "muto/org.eclipse.muto.sandbox:f1tenth",
-                    "correlation-id": "7c4a0d1f-c38a-4076-b138-ac07357c1d0e"
+                    "correlation-id": "7c4a0d1f-c38a-4076-b138-ac07357c1d0e",
                 },
                 "path": "/inbox/messages/agent/commands/ping",
-                "value": {}
+                "value": {},
             }
         )
         gw_msg.meta = meta_msg
@@ -80,7 +73,7 @@ class TestAgentNode(unittest.TestCase):
             Gateway,
             self.node.get_parameter("agent_to_gateway_topic").value,
             lambda msg: setattr(self.node, "received_message", msg),
-            10
+            10,
         )
 
         self.node._gateway_msg_callback(gw_msg)
@@ -105,10 +98,10 @@ class TestAgentNode(unittest.TestCase):
                 "headers": {
                     "content-type": "application/json",
                     "reply-to": "muto/org.eclipse.muto.sandbox:alp-001",
-                    "correlation-id": "082e0be4-83f3-4a71-a287-aa242ef3bc91"
+                    "correlation-id": "082e0be4-83f3-4a71-a287-aa242ef3bc91",
                 },
                 "path": "/inbox/messages/stack/commands/kill",
-                "value": {"stackId": "org.eclipse.muto.sandbox:composable_client_server"}
+                "value": {"stackId": "org.eclipse.muto.sandbox:composable_client_server"},
             }
         )
 
@@ -119,7 +112,7 @@ class TestAgentNode(unittest.TestCase):
             MutoAction,
             self.node.get_parameter("stack_topic").value,
             lambda msg: setattr(self.node, "received_message", msg),
-            10
+            10,
         )
 
         self.node._gateway_msg_callback(gw_msg)
@@ -138,17 +131,19 @@ class TestAgentNode(unittest.TestCase):
         meta_msg.correlation_data = "494da5fa-4204-4b2e-8d10-d3138af3349a"
 
         gw_msg = Gateway()
-        gw_msg.topic = "org.eclipse.muto.sandbox/f1tenth/things/live/messages/agent/commands/ros/node"
+        gw_msg.topic = (
+            "org.eclipse.muto.sandbox/f1tenth/things/live/messages/agent/commands/ros/node"
+        )
         gw_msg.payload = json.dumps(
             {
                 "topic": "org.eclipse.muto.sandbox/f1tenth/things/live/messages/agent/commands/ros/node",
                 "headers": {
                     "content-type": "application/json",
                     "reply-to": "muto/org.eclipse.muto.sandbox:f1tenth",
-                    "correlation-id": "494da5fa-4204-4b2e-8d10-d3138af3349a"
+                    "correlation-id": "494da5fa-4204-4b2e-8d10-d3138af3349a",
                 },
                 "path": "/inbox/messages/agent/commands/ros/node",
-                "value": {}
+                "value": {},
             }
         )
         gw_msg.meta = meta_msg
@@ -158,7 +153,7 @@ class TestAgentNode(unittest.TestCase):
             MutoAction,
             self.node.get_parameter("agent_to_commands_topic").value,
             lambda msg: setattr(self.node, "received_message", msg),
-            10
+            10,
         )
 
         self.node._gateway_msg_callback(gw_msg)
@@ -177,9 +172,11 @@ class TestAgentNode(unittest.TestCase):
 
     def test_commands_msg_callback(self):
         meta_msg = MutoActionMeta()
-        meta_msg.response_topic = "db-org.eclipse.muto.sandbox:f1tenth/agent/fa30fe44-153f-4541-b78e-3f71863a331d"
-        meta_msg.correlation_data = "fa30fe44-153f-4541-b78e-3f71863a331d"        
-        
+        meta_msg.response_topic = (
+            "db-org.eclipse.muto.sandbox:f1tenth/agent/fa30fe44-153f-4541-b78e-3f71863a331d"
+        )
+        meta_msg.correlation_data = "fa30fe44-153f-4541-b78e-3f71863a331d"
+
         action_msg = MutoAction()
         action_msg.context = ""
         action_msg.method = ""
@@ -191,7 +188,7 @@ class TestAgentNode(unittest.TestCase):
             Gateway,
             self.node.get_parameter("agent_to_gateway_topic").value,
             lambda msg: setattr(self.node, "received_message", msg),
-            10
+            10,
         )
 
         self.node._commands_msg_callback(action_msg)
@@ -205,7 +202,7 @@ class TestAgentNode(unittest.TestCase):
 
     def test_parse_topic(self):
         topic = "org.eclipse.muto.sandbox/f1tenth/things/live/messages/agent/commands/ping"
-        
+
         # Use the topic parser from the node
         topic_parser = self.node.get_topic_parser()
         if topic_parser:
@@ -214,11 +211,11 @@ class TestAgentNode(unittest.TestCase):
         else:
             # If topic parser is not available, skip the test
             self.skipTest("Topic parser not initialized")
-        
+
         topic = "org.eclipse.muto.sandbox:f1tenth/stack/commands/apply"
         res = self.node.parse_topic(topic)
         assert res == ("stack", "apply"), "Return value must be ('stack', 'apply')"
-        
+
         topic = "org.eclipse.muto.sandbox/f1tenth/things/live/messages/agent/commands/ros/node"
         res = self.node.parse_topic(topic)
         assert res == ("agent", "ros/node"), "Return value must be ('agent', 'ros/node')"
