@@ -4,23 +4,23 @@ Comprehensive unit tests for Symphony SDK summary models.
 """
 
 import unittest
-import json
+
 from symphony_sdk.summary import (
-    SummaryState,
     ComponentResultSpec,
-    TargetResultSpec,
     SummaryResult,
     SummarySpec,
-    create_success_component_result,
+    SummaryState,
+    TargetResultSpec,
     create_failed_component_result,
-    create_target_result
+    create_success_component_result,
+    create_target_result,
 )
 from symphony_sdk.types import State
 
 
 class TestSummaryState(unittest.TestCase):
     """Test cases for SummaryState enumeration."""
-    
+
     def test_summary_state_values(self):
         """Test SummaryState enum values."""
         self.assertEqual(SummaryState.PENDING, 0)
@@ -36,7 +36,7 @@ class TestSummaryState(unittest.TestCase):
 
 class TestComponentResultSpec(unittest.TestCase):
     """Test cases for ComponentResultSpec."""
-    
+
     def test_component_result_spec_creation(self):
         """Test ComponentResultSpec creation with default values."""
         result = ComponentResultSpec()
@@ -45,34 +45,22 @@ class TestComponentResultSpec(unittest.TestCase):
 
     def test_component_result_spec_with_values(self):
         """Test ComponentResultSpec creation with specific values."""
-        result = ComponentResultSpec(
-            status=State.UPDATE_FAILED,
-            message="Update operation failed"
-        )
+        result = ComponentResultSpec(status=State.UPDATE_FAILED, message="Update operation failed")
         self.assertEqual(result.status, State.UPDATE_FAILED)
         self.assertEqual(result.message, "Update operation failed")
 
     def test_component_result_spec_to_dict(self):
         """Test ComponentResultSpec to_dict conversion."""
-        result = ComponentResultSpec(
-            status=State.UPDATED,
-            message="Component updated successfully"
-        )
-        
-        expected_dict = {
-            "status": State.UPDATED.value,
-            "message": "Component updated successfully"
-        }
-        
+        result = ComponentResultSpec(status=State.UPDATED, message="Component updated successfully")
+
+        expected_dict = {"status": State.UPDATED.value, "message": "Component updated successfully"}
+
         self.assertEqual(result.to_dict(), expected_dict)
 
     def test_component_result_spec_from_dict(self):
         """Test ComponentResultSpec from_dict conversion."""
-        data = {
-            "status": State.DELETE_FAILED.value,
-            "message": "Failed to delete component"
-        }
-        
+        data = {"status": State.DELETE_FAILED.value, "message": "Failed to delete component"}
+
         result = ComponentResultSpec.from_dict(data)
         self.assertEqual(result.status, State.DELETE_FAILED)
         self.assertEqual(result.message, "Failed to delete component")
@@ -94,7 +82,7 @@ class TestComponentResultSpec(unittest.TestCase):
 
 class TestTargetResultSpec(unittest.TestCase):
     """Test cases for TargetResultSpec."""
-    
+
     def test_target_result_spec_creation(self):
         """Test TargetResultSpec creation with defaults."""
         result = TargetResultSpec()
@@ -106,16 +94,13 @@ class TestTargetResultSpec(unittest.TestCase):
         """Test TargetResultSpec creation with specific values."""
         comp1 = ComponentResultSpec(State.UPDATED, "Component 1 updated")
         comp2 = ComponentResultSpec(State.UPDATE_FAILED, "Component 2 failed")
-        
+
         result = TargetResultSpec(
             status="PARTIAL_SUCCESS",
             message="Some components failed",
-            component_results={
-                "comp1": comp1,
-                "comp2": comp2
-            }
+            component_results={"comp1": comp1, "comp2": comp2},
         )
-        
+
         self.assertEqual(result.status, "PARTIAL_SUCCESS")
         self.assertEqual(result.message, "Some components failed")
         self.assertEqual(len(result.component_results), 2)
@@ -128,29 +113,24 @@ class TestTargetResultSpec(unittest.TestCase):
         result = TargetResultSpec(
             status="SUCCESS",
             message="All components updated",
-            component_results={"comp1": comp_result}
+            component_results={"comp1": comp_result},
         )
-        
+
         result_dict = result.to_dict()
-        
+
         expected_dict = {
             "status": "SUCCESS",
             "message": "All components updated",
-            "components": {
-                "comp1": {
-                    "status": State.UPDATED.value,
-                    "message": "Success"
-                }
-            }
+            "components": {"comp1": {"status": State.UPDATED.value, "message": "Success"}},
         }
-        
+
         self.assertEqual(result_dict, expected_dict)
 
     def test_target_result_spec_to_dict_minimal(self):
         """Test TargetResultSpec to_dict with minimal data."""
         result = TargetResultSpec(status="OK")
         result_dict = result.to_dict()
-        
+
         expected_dict = {"status": "OK"}
         self.assertEqual(result_dict, expected_dict)
 
@@ -160,30 +140,24 @@ class TestTargetResultSpec(unittest.TestCase):
             "status": "FAILED",
             "message": "Operation failed",
             "components": {
-                "comp1": {
-                    "status": State.UPDATE_FAILED.value,
-                    "message": "Component update failed"
-                }
-            }
+                "comp1": {"status": State.UPDATE_FAILED.value, "message": "Component update failed"}
+            },
         }
-        
+
         result = TargetResultSpec.from_dict(data)
-        
+
         self.assertEqual(result.status, "FAILED")
         self.assertEqual(result.message, "Operation failed")
         self.assertEqual(len(result.component_results), 1)
-        
+
         comp_result = result.component_results["comp1"]
         self.assertEqual(comp_result.status, State.UPDATE_FAILED)
         self.assertEqual(comp_result.message, "Component update failed")
 
     def test_target_result_spec_from_dict_no_components(self):
         """Test TargetResultSpec from_dict without components."""
-        data = {
-            "status": "SUCCESS",
-            "message": "No components to process"
-        }
-        
+        data = {"status": "SUCCESS", "message": "No components to process"}
+
         result = TargetResultSpec.from_dict(data)
         self.assertEqual(result.status, "SUCCESS")
         self.assertEqual(result.message, "No components to process")
@@ -192,16 +166,12 @@ class TestTargetResultSpec(unittest.TestCase):
 
 class TestSummaryResult(unittest.TestCase):
     """Test cases for SummaryResult."""
-    
+
     def test_summary_result_creation(self):
         """Test SummaryResult creation."""
         summary_spec = SummarySpec()
-        summary = SummaryResult(
-            summary=summary_spec,
-            state=SummaryState.DONE,
-            generation="1"
-        )
-        
+        summary = SummaryResult(summary=summary_spec, state=SummaryState.DONE, generation="1")
+
         self.assertEqual(summary.state, SummaryState.DONE)
         self.assertEqual(summary.generation, "1")
         self.assertEqual(summary.summary, summary_spec)
@@ -211,7 +181,7 @@ class TestSummaryResult(unittest.TestCase):
         # Test with DONE state
         summary_done = SummaryResult(state=SummaryState.DONE)
         self.assertTrue(summary_done.is_deployment_finished())
-        
+
         # Test with RUNNING state
         summary_running = SummaryResult(state=SummaryState.RUNNING)
         self.assertFalse(summary_running.is_deployment_finished())
@@ -219,14 +189,11 @@ class TestSummaryResult(unittest.TestCase):
 
 class TestSummarySpec(unittest.TestCase):
     """Test cases for SummarySpec."""
-    
+
     def test_summary_spec_creation(self):
         """Test SummarySpec creation."""
-        spec = SummarySpec(
-            target_count=2,
-            success_count=1
-        )
-        
+        spec = SummarySpec(target_count=2, success_count=1)
+
         self.assertEqual(spec.target_count, 2)
         self.assertEqual(spec.success_count, 1)
 
@@ -239,32 +206,32 @@ class TestSummarySpec(unittest.TestCase):
 
 class TestHelperFunctions(unittest.TestCase):
     """Test cases for helper functions."""
-    
+
     def test_create_success_component_result(self):
         """Test create_success_component_result function."""
         result = create_success_component_result("Operation completed")
-        
+
         self.assertEqual(result.status, State.OK)
         self.assertEqual(result.message, "Operation completed")
 
     def test_create_success_component_result_default_message(self):
         """Test create_success_component_result with default message."""
         result = create_success_component_result()
-        
+
         self.assertEqual(result.status, State.OK)
         self.assertEqual(result.message, "")
 
     def test_create_failed_component_result(self):
         """Test create_failed_component_result function."""
         result = create_failed_component_result("Operation failed", State.UPDATE_FAILED)
-        
+
         self.assertEqual(result.status, State.UPDATE_FAILED)
         self.assertEqual(result.message, "Operation failed")
 
     def test_create_failed_component_result_default_state(self):
         """Test create_failed_component_result with default state."""
         result = create_failed_component_result("Something went wrong")
-        
+
         self.assertEqual(result.status, State.INTERNAL_ERROR)
         self.assertEqual(result.message, "Something went wrong")
 
@@ -272,11 +239,11 @@ class TestHelperFunctions(unittest.TestCase):
         """Test create_target_result function."""
         components = {
             "comp1": create_success_component_result("Updated"),
-            "comp2": create_failed_component_result("Failed")
+            "comp2": create_failed_component_result("Failed"),
         }
-        
+
         result = create_target_result("PARTIAL", "Some components failed", components)
-        
+
         self.assertEqual(result.status, "PARTIAL")
         self.assertEqual(result.message, "Some components failed")
         self.assertEqual(len(result.component_results), 2)
@@ -286,11 +253,11 @@ class TestHelperFunctions(unittest.TestCase):
     def test_create_target_result_minimal(self):
         """Test create_target_result with minimal parameters."""
         result = create_target_result()
-        
+
         self.assertEqual(result.status, "OK")
         self.assertEqual(result.message, "")
         self.assertEqual(result.component_results, {})
 
 
-if __name__ == '__main__':
+if __name__ == "__main__":
     unittest.main()
